refactor: minor changes menulist & style of main content
