+++ frontend/src/components/Navbar/DesktopMenu.jsx
@@ -1,4 +1,5 @@
 import React from "react";
+import { NavLink } from "react-router-dom";
 import { getFilteredMenuList, menuList } from "../../utils";
 
 function DesktopMenu({ token, handleSideMenuClick }) {
