+++ frontend/src/styles.css
@@ -17,7 +17,7 @@ body {
   left: 0;
   right: 0;
   z-index: var(--z-index-9);
-  padding: 0rem 0.5rem;
+  padding: 0.5rem;
   align-items: center;
   box-shadow: 0 2px 5px rgb(0 0 0 / 20%);
   border-radius: 10px;
@@ -54,7 +54,7 @@ body {
 }
 
 .main-content {
-  margin-top: 5rem;
+  margin-top: 4rem;
 }
 
 .side-bar {
