+++ frontend/src/utils/loginUtils.js
@@ -34,3 +34,11 @@ export const getFormValues = (e, type) => {
 export const getFilteredMenuList = (menuList, token) => {
   return token ? menuList.filter(({ name }) => name !== "Login") : menuList;
 };
+
+export const menuList = [
+  { name: "Home", icon: "fa-home-alt", path: "/" },
+  { name: "Products", icon: "fa-list-ul", path: "/products" },
+  { name: "Wishlist", icon: "fa-heart", path: "/wish" },
+  { name: "Cart", icon: "fa-shopping-cart", path: "/cart" },
+  { name: "Login", icon: "fa-sign-in-alt", path: "/login" },
+];
